+++ app-backend/backsplash/src/main/scala/services/MosaicService.scala
@@ -4,6 +4,7 @@ import com.azavea.maml.error.Interpreted
 import com.azavea.maml.eval.BufferingInterpreter
 import com.rasterfoundry.authentication.Authentication
 import com.rasterfoundry.backsplash.error._
+import com.rasterfoundry.backsplash.io.Histogram
 import com.rasterfoundry.backsplash.nodes.ProjectNode
 import com.rasterfoundry.backsplash.parameters.PathParameters._
 import com.rasterfoundry.common.RollbarNotifier
@@ -18,9 +19,22 @@ import cats.implicits._
 import doobie.implicits._
 import geotrellis.raster.Tile
 import geotrellis.server.core.maml._
+import geotrellis.server.core.maml.reification.MamlTmsReification
+import geotrellis.vector.{Projected, Polygon}
+import io.circe._
+import io.circe.syntax._
+import io.circe.parser._
 import org.http4s._
 import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
+import org.http4s.implicits._
+import org.http4s.server._
+import org.http4s.util.CaseInsensitiveString
+
+import java.net.URI
+import java.util.UUID
+import java.util.Base64
 
 class MosaicService(
     interpreter: BufferingInterpreter = BufferingInterpreter.DEFAULT
@@ -97,5 +111,46 @@ class MosaicService(
           }
         } yield resp
         respIO.handleErrorWith(handleErrors _)
+
+      case GET -> Root / UUIDWrapper(projectId) / "histogram" / _
+            :? RedBandOptionalQueryParamMatcher(redOverride)
+            :? GreenBandOptionalQueryParamMatcher(greenOverride)
+            :? BlueBandOptionalQueryParamMatcher(blueOverride) as user =>
+        val respIO = for {
+          histograms <- Histogram.getRGBProjectHistogram(projectId,
+                                                         None,
+                                                         None,
+                                                         None,
+                                                         None,
+                                                         List.empty[UUID])
+          resp <- Ok((histograms map { _.binCounts.toMap } asJson).noSpaces)
+        } yield resp
+        respIO handleErrorWith { handleErrors _ }
+
+      case authedReq @ POST -> Root / UUIDWrapper(projectId) / "histogram" / _
+            :? RedBandOptionalQueryParamMatcher(redOverride)
+            :? GreenBandOptionalQueryParamMatcher(greenOverride)
+            :? BlueBandOptionalQueryParamMatcher(blueOverride) as user => {
+        // Compile to a byte array, decode that as a string, and do something with the results
+        authedReq.req.body.compile.to[Array] flatMap { uuids =>
+          decode[List[UUID]](
+            uuids map { _.toChar } mkString
+          ) match {
+            case Right(uuids) =>
+              logger.info(s"How many uuids: ${uuids.length}")
+              for {
+                histograms <- Histogram.getRGBProjectHistogram(projectId,
+                                                               None,
+                                                               redOverride,
+                                                               greenOverride,
+                                                               blueOverride,
+                                                               uuids)
+                resp <- Ok(
+                  (histograms map { _.binCounts.toMap } asJson).noSpaces)
+              } yield resp
+            case _ => BadRequest("Could not decode body as sequence of UUIDs")
+          }
+        } handleErrorWith (handleErrors _)
+      }
     }
 }